requirements.txt
agno==1.5.9
anthropic==0.52.2
click==8.2.1
cohere
Django==5.0.1
django_cors_headers==4.4.0
djangorestframework==3.15.2
e2b==1.5.1
spider-client
googlesearch-python
pycountry
invoke==2.2.0
ollama
infinity_client
sqlalchemy
google-genai
lancedb
yfinance==0.2.54
openai==1.84.0
protobuf==5.29.4
python-dotenv==1.1.0
Requests==2.32.3
sseclient_py==1.8.0
uvicorn
watchdog==4.0.1
duckduckgo-search
